+++ CHANGES.md
@@ -1,4 +1,6 @@
 
+* Add an `Mtime.t` type equal to `Mtime.span`.
+* Add `Mtime.{compare,equal}. Thanks to David Sheets for the patch.
 * Add `Mtime.of_ns_uint64`. Thanks to David Sheets for the patch.
 
 v0.8.4 2017-02-05 La Forclaz (VS)